Spring Security added(Authentication and Authorization), db triggers realized in controllers, additional bugfixing and testing needed

+++ pcpartsbackend/src/main/java/erp/pcpartsbackend/models/Order.java
@@ -25,7 +25,7 @@ public class Order implements Serializable {
     private Date orderDate;
     private Float orderPrice;
     private String orderStatus;
-    private Float discount;
+    private double discount;
     private String promoCode;
     @JsonIgnore
     @ManyToOne(fetch = FetchType.LAZY)
